增加了回溯对话和替角色说功能。

+++ modules/methods.py
@@ -52,7 +52,7 @@ def characters_get():
   user_name = flask.request.values.get('user_name')
   character_name = flask.request.values.get('character_name')
   if not character_name:
-    return return_error('缺少关键参数')
+    return return_error('关键参数为空')
   character_path = f'./chars/{user_name}/{character_name}.json'
   if not os.path.exists(character_path):
     return return_error('角色不存在') 
@@ -64,7 +64,6 @@ def characters_get():
 @methods.route("/characters/save", methods=['post'])
 def characters_save():
   user_name = flask.request.values.get('user_name')
-  bot_save_name = flask.request.values.get('bot_save_name')
   user = flask.request.values.get('user')
   bot = flask.request.values.get('bot')
   action_start = flask.request.values.get('action_start')
@@ -72,16 +71,11 @@ def characters_save():
   greeting = flask.request.values.get('greeting')
   bot_persona = flask.request.values.get('bot_persona')
   example_message = flask.request.values.get('example_message')
-  use_qa = flask.request.values.get('use_qa', False)
+  use_qa = flask.request.values.get('use_qa', False, type=bool)
   avatar = flask.request.values.get('avatar')
   if not user or not bot or not greeting or not bot_persona:
-    return return_error('缺少关键参数')
-  bot_name = bot if not bot_save_name else bot_save_name
-  with open(f"./chars/{user_name}/{bot_name}.json", 'w', encoding='utf8') as f:
-    if not use_qa or use_qa.lower() == 'false':
-      use_qa = False
-    else:
-      use_qa = True
+    return return_error('关键参数为空')
+  with open(f"./chars/{user_name}/{bot}.json", 'w', encoding='utf8') as f:
     data = {
       'user': user, 
       'bot': bot, 
@@ -94,6 +88,9 @@ def characters_save():
       'avatar': avatar
     }
     json.dump(data, f, indent=2, ensure_ascii=False)
+  sav_path = f"save/{user_name}/{bot}.sav"
+  if os.path.exists(sav_path):
+    os.remove(sav_path)
   return return_success()
 
 # 删除角色
@@ -101,15 +98,12 @@ def characters_save():
 def characters_delete():
   user_name = flask.request.values.get('user_name')
   char_name = flask.request.values.get('character_name')
-  if not user_name or not char_name:
-    return return_error('缺少关键参数')
+  if not char_name:
+    return return_error('关键参数为空')
   json_path = f"chars/{user_name}/{char_name}.json"
-  pic_path = f"chars/{user_name}/{char_name}.png"
   sav_path = f"save/{user_name}/{char_name}.sav"
   if os.path.exists(json_path):
     os.remove(json_path)
-  if os.path.exists(pic_path):
-    os.remove(pic_path)
   if os.path.exists(sav_path):
     os.remove(sav_path)
   return return_success()
@@ -122,6 +116,8 @@ def characters_load():
   if not char_name:
     return return_error('关键参数为空')
   role_info = init_chat(user_name, char_name)
+  if not role_info:
+    return return_error('角色不存在')
   data = {
     'chat': role_info.chatbot
   }
@@ -130,7 +126,7 @@ def characters_load():
 def init_chat(user_name, char_name):
   char_path = f'./chars/{user_name}/{char_name}.json'
   if not os.path.exists(char_path):
-    return return_error('角色不存在')
+    return False
   with open(char_path, 'r', encoding='utf-8') as f:
     char = json.loads(f.read())
   role_info = RoleInfo([], char['user'], char['bot'], char['action_start'], char['action_end'], 
@@ -139,14 +135,6 @@ def init_chat(user_name, char_name):
   model_tokens = []
   model_state = None
   init_prompt = get_init_prompt(role_info, char['bot'], char['bot_persona'], char['user'], char['example_message'])
-  init_prompt = init_prompt.strip().split('\n')
-  for c in range(len(init_prompt)):
-    init_prompt[c] = init_prompt[c].strip().strip('\u3000').strip('\r')
-  init_prompt = '\n'.join(init_prompt).strip() + '\n\n'
-  init_prompt = init_prompt.strip().split('\n')
-  for c in range(len(init_prompt)):
-    init_prompt[c] = init_prompt[c].strip().strip('\u3000').strip('\r')
-  init_prompt = '\n'.join(init_prompt).strip() + '\n\n'
   if greeting:
     init_prompt += f"{role_info.bot}: {greeting}\n\n"
     role_info.chatbot = [[None, greeting]]
@@ -154,7 +142,9 @@ def init_chat(user_name, char_name):
     out, model_tokens, model_state = model.run_rnn(model_tokens, model_state, model.pipeline.encode(init_prompt))
     save_state(user_name, role_info, out, model_tokens, model_state)
   else:
-    save_data = load_state(user_name, role_info.bot)
+    save_data = load_state(user_name, role_info.bot_chat)
+    if not save_data:
+      return False
     role_info = save_data['role_info']
   return role_info
 
@@ -164,12 +154,12 @@ def chat_reply():
   user_name = flask.request.values.get('user_name')
   char_name = flask.request.values.get('character_name')
   prompt = flask.request.values.get('prompt')
-  min_len = flask.request.values.get('min_len', 0)
-  top_p = flask.request.values.get('top_p', 0.65)
-  temperature = flask.request.values.get('temperature', 2)
-  presence_penalty = flask.request.values.get('presence_penalty', 0.2)
-  frequency_penalty = flask.request.values.get('frequency_penalty', 0.2)
-  if not user_name or not prompt or not char_name:
+  min_len = flask.request.values.get('min_len', 0, type=int)
+  top_p = flask.request.values.get('top_p', 0.65, type=float)
+  temperature = flask.request.values.get('temperature', 2, type=float)
+  presence_penalty = flask.request.values.get('presence_penalty', 0.2, type=float)
+  frequency_penalty = flask.request.values.get('frequency_penalty', 0.2, type=float)
+  if not prompt or not char_name:
     return return_error('关键参数为空')
   save_data = load_state(user_name, char_name)
   if not save_data:
@@ -192,11 +182,13 @@ def chat_reply():
 def chat_resay():
   user_name = flask.request.values.get('user_name')
   char_name = flask.request.values.get('character_name')
-  min_len = flask.request.values.get('min_len', 0)
-  top_p = flask.request.values.get('top_p', 0.65)
-  temperature = flask.request.values.get('temperature', 2)
-  presence_penalty = flask.request.values.get('presence_penalty', 0.2)
-  frequency_penalty = flask.request.values.get('frequency_penalty', 0.2)
+  min_len = flask.request.values.get('min_len', 0, type=int)
+  top_p = flask.request.values.get('top_p', 0.65, type=float)
+  temperature = flask.request.values.get('temperature', 2, type=float)
+  presence_penalty = flask.request.values.get('presence_penalty', 0.2, type=float)
+  frequency_penalty = flask.request.values.get('frequency_penalty', 0.2, type=float)
+  if not char_name:
+    return return_error('关键参数为空')
   save_data = load_state(user_name, char_name)
   if not save_data:
     return return_error('尚未加载角色')
@@ -217,10 +209,14 @@ def chat_resay():
 def chat_reset():
   user_name = flask.request.values.get('user_name')
   char_name = flask.request.values.get('character_name')
+  if not char_name:
+    return return_error('关键参数为空')
   save_file = f"save/{user_name}/{char_name}.sav"
   if os.path.exists(save_file):
     os.remove(save_file)
   role_info = init_chat(user_name, char_name)
+  if not role_info:
+    return return_error('角色不存在')
   data = {
     'chat': role_info.chatbot
   }
@@ -231,6 +227,8 @@ def chat_reset():
 def debug_token():
   user_name = flask.request.values.get('user_name')
   char_name = flask.request.values.get('character_name')
+  if not char_name:
+    return return_error('关键参数为空')
   save_data = load_state(user_name, char_name)
   if not save_data:
     return return_error('尚未开始对话')
@@ -239,6 +237,69 @@ def debug_token():
     'token_state': model.pipeline.decode(save_data['model_tokens'])
   }
   return return_success(data)
+
+# 回溯对话
+@methods.route("/chat/back", methods=['post'])
+def chat_back():
+  user_name = flask.request.values.get('user_name')
+  char_name = flask.request.values.get('character_name')
+  log_index = flask.request.values.get('log_index', 0, type=int)
+  # 先把需要的数据读出来
+  if not char_name:
+    return return_error('关键参数为空')
+  if not log_index or log_index < 0:
+    log_index = 0
+  char_path = f'./chars/{user_name}/{char_name}.json'
+  if not os.path.exists(char_path):
+    return return_error('角色不存在')
+  with open(char_path, 'r', encoding='utf-8') as f:
+    char = json.loads(f.read())
+  save_data = load_state(user_name, char_name)
+  role_info = save_data['role_info']
+  chatbot = role_info.chatbot
+  # 根据log_index来截取对话，然后从头重新构建chat_pre和chat
+  init_prompt = get_init_prompt(save_data['role_info'], char['bot'], char['bot_persona'], char['user'], char['example_message'])
+  chatbot_pre = chatbot[0:log_index]
+  next_reply = chatbot[log_index][1]
+  role_info.chatbot = chatbot[0:log_index + 1]
+  for row in chatbot_pre:
+    if row[0]:
+      init_prompt += f"{role_info.user}: {row[0]}\n\n"
+    if row[1]:
+      init_prompt += f"{role_info.bot}: {row[1]}\n\n"
+  model_tokens = []
+  model_state = None
+  out_pre, model_tokens_pre, model_state_pre = model.run_rnn(model_tokens, model_state, model.pipeline.encode(init_prompt))
+  mtp = copy.deepcopy(model_tokens_pre)
+  msp = copy.deepcopy(model_state_pre)
+  new = f"{save_data['role_info'].bot}: {next_reply}\n\n"
+  out, model_tokens, model_state = model.run_rnn(model_tokens_pre, model_state_pre, model.pipeline.encode(new))
+  # 当log_index为0时，需要特殊处理
+  if log_index == 0:
+    save_state(user_name, role_info, out, model_tokens, model_state, None, None, None)
+  else:
+    save_state(user_name, role_info, out, model_tokens, model_state, out_pre, mtp, msp)
+  return return_success()
+
+# 替角色说
+@methods.route("/chat/tamper", methods=['post'])
+def chat_tamper():
+  user_name = flask.request.values.get('user_name')
+  char_name = flask.request.values.get('character_name')
+  message = flask.request.values.get('message')
+  if not char_name or not message:
+    return return_error('关键参数为空')
+  save_data = load_state(user_name, char_name)
+  if not save_data:
+    return return_error('尚未加载角色')
+  if not save_data['model_tokens_pre']:
+    return return_error('尚未开始对话')
+  role_info = save_data['role_info']
+  role_info.chatbot[-1][1] = message
+  new = f" {message}\n\n"
+  out, model_tokens, model_state = model.run_rnn(copy.deepcopy(save_data['model_tokens_pre']), copy.deepcopy(save_data['model_state_pre']), model.pipeline.encode(new))
+  save_state(user_name, role_info, out, model_tokens, model_state, save_data['out_pre'], save_data['model_tokens_pre'], save_data['model_state_pre'])
+  return return_success()
 
 def gen_msg(chat_param, out_pre, model_tokens_pre, model_state_pre, user_name, role_info:RoleInfo, occurrence):
   c_model_tokens_pre = copy.deepcopy(model_tokens_pre)
@@ -264,6 +325,14 @@ def get_init_prompt(role_info:RoleInfo, bot:str, bot_persona:str, user:str, exam
   else:
     init_prompt = f'{init_prompt_part2}'
   init_prompt += f"{bot_persona}"
+  init_prompt = init_prompt.strip().split('\n')
+  for c in range(len(init_prompt)):
+    init_prompt[c] = init_prompt[c].strip().strip('\u3000').strip('\r')
+  init_prompt = '\n'.join(init_prompt).strip() + '\n\n'
+  init_prompt = init_prompt.strip().split('\n')
+  for c in range(len(init_prompt)):
+    init_prompt[c] = init_prompt[c].strip().strip('\u3000').strip('\r')
+  init_prompt = '\n'.join(init_prompt).strip() + '\n\n'
   return init_prompt
 
 def save_state(user_name, role_info:RoleInfo, out, model_tokens, model_state, out_pre=None, model_tokens_pre=None, model_state_pre=None):
